arm64: dts: allwinner: pinephone: Add mount-matrix for LIS3MDL magnetometer
The way STMicroelectronics LIS3MDL is mounted on the user-facing side
of the Pine64 PinePhone mainboard requires the magnetometers's x- and
y-axis to be swapped, and the direction of the accelerometer's x-axis
to be inverted.

Rectify this by adding a mount-matrix to the magnetometer's
definition in the PinePhone dtsi file.

Signed-off-by: Shoji Keita <awaittrot@shjk.jp>
Signed-off-by: Andrey Skvortsov <andrej.skvortzov@gmail.com>

Gbp-Pq: Topic magnetometer
Gbp-Pq: Name 0088-arm64-dts-sun50i-a64-pinephone-Add-mount-matrix-for-.patch