arm64: dts: allwinner: pinephone: Add mount-matrix for AF8133J magnetometer
The way Voltafield AF8133J is mounted on the user-facing side of the
Pine64 PinePhone mainboard requires the magnetometers's x- and y-axis
to be swapped, and direction of all axis to be inverted.

Rectify this by adding a mount-matrix to the magnetometer's
definition in the PinePhone dtsi file.

Signed-off-by: Shoji Keita <awaittrot@shjk.jp>
Signed-off-by: Andrey Skvortsov <andrej.skvortzov@gmail.com>

Gbp-Pq: Topic magnetometer
Gbp-Pq: Name 0087-arm64-dts-allwinner-pinephone-Add-mount-matrix-for-A.patch